feat: :rocket: Exercise 10
Develop a cyclical program that captures a piece of data number each time, and accumulate them. The program will      stops when the user types a zero. The program must      show: THE SUM OF THE VALUES, THE VALUE OF THE AVERAGE, HOW MANY VALUES WERE ENTERED, MAYOR VALUE AND LESS VALUE